Setup folder structure for backend
This chore branch sets up the bckend structure to organize files
into routes and controllers. Two directories, routes and controllers
have been created and each has files to hold backend logic.
The files just create empty functions for now.

In addition to this, there is a slight change in the database to have
separate tables for gifs and article comments.
This is reflected in the database-schema.sql file.